requirements.txt
djangorestframework 
djangorestframework-simplejwt
django-cors-headers
django
requests
